More function caching
Removed functions as arguments, which makes the code less modular, but allows caching all the numba compiled functions
