+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "hatchling.build"
 [project]
 name = "foostrap"
 description = "Fast Bootstrap statistics using Numba"
-version = "1.1.0"
+version = "1.1.1"
 dependencies = [
   "numpy>=1.25",
   "numba>=0.58",
